fix: 리뷰 반영 - !== → !=, currentCount Int → Long

+++ SClass-Api-Supporters/src/main/kotlin/com/sclass/supporters/enrollment/usecase/PrepareEnrollmentUseCase.kt
@@ -36,7 +36,7 @@ class PrepareEnrollmentUseCase(
     ): PrepareEnrollmentResponse {
         val course = courseAdaptor.findById(courseId)
         val liveCount = enrollmentAdaptor.countLiveEnrollments(courseId)
-        if (!course.canEnroll(LocalDateTime.now(), liveCount.toInt())) {
+        if (!course.canEnroll(LocalDateTime.now(), liveCount)) {
             throw CourseNotEnrollableException()
         }
